move Core::Config into Util::Config
