+++ lib/munge.rb
@@ -9,6 +9,7 @@ require "munge/version"
 require "munge/item"
 require "munge/util/path"
 require "munge/util/symbol_hash"
+require "munge/util/config"
 require "munge/helper"
 require "munge/helper/find"
 require "munge/helper/link"
@@ -18,7 +19,6 @@ require "munge/readers/filesystem"
 require "munge/routers/auto_add_extension"
 require "munge/routers/fingerprint"
 require "munge/routers/index_html"
-require "munge/core/config"
 require "munge/core/router"
 require "munge/core/item_factory"
 require "munge/core/collection"
