+++ lib/munge/util/config.rb
